sdm845-common: sepolicy: Remove undesired camera prop label
This was basically forcing all props starting with vendor.camera. to
be labeled as camera_prop.

The side effect of doing that was the relabel of props that were
supposed to have a particular label for some reason. Take for example
vendor.camera.aux.packagelist which isn't desired to be relabeled
so to avoid having to write additional local sepolicies.

Change-Id: I1b7c952a241688f6517e1742719e77b6141e2cd2
Signed-off-by: Akhil Narang <me@akhilnarang.dev>